requirements.txt
django
psycopg2-binary
django-environ
django-widget-tweaks
google-api-python-client